+++ wpull/util.py
@@ -3,9 +3,7 @@
 from bs4.dammit import UnicodeDammit, EncodingDetector
 import calendar
 import codecs
-import collections
 import contextlib
-import copy
 import datetime
 import itertools
 import os.path
@@ -18,57 +16,6 @@ import tornado.util
 import toro
 import zipfile
 import zlib
-
-
-try:
-    from collections import OrderedDict
-except ImportError:
-    from wpull.backport.collections import OrderedDict
-
-
-class OrderedDefaultDict(OrderedDict):
-    '''An ordered default dict.
-
-    http://stackoverflow.com/a/6190500/1524507
-    '''
-    def __init__(self, default_factory=None, *args, **kwargs):
-        if default_factory is not None and \
-        not isinstance(default_factory, collections.Callable):
-            raise TypeError('First argument must be callable')
-        OrderedDict.__init__(self, *args, **kwargs)
-        self.default_factory = default_factory
-
-    def __getitem__(self, key):
-        try:
-            return OrderedDict.__getitem__(self, key)
-        except KeyError:
-            return self.__missing__(key)
-
-    def __missing__(self, key):
-        if self.default_factory is None:
-            raise KeyError(key)
-        self[key] = value = self.default_factory()
-        return value
-
-    def __reduce__(self):
-        if self.default_factory is None:
-            args = tuple()
-        else:
-            args = self.default_factory,
-        return type(self), args, None, None, self.items()
-
-    def copy(self):
-        return self.__copy__()
-
-    def __copy__(self):
-        return type(self)(self.default_factory, self)
-
-    def __deepcopy__(self, memo):
-        return type(self)(self.default_factory, copy.deepcopy(self.items()))
-
-    def __repr__(self):
-        return 'OrderedDefaultDict(%s, %s)' % (
-            self.default_factory, collections.OrderedDict.__repr__(self))
 
 
 class ASCIIStreamWriter(codecs.StreamWriter):